Исправил тесты и прочие недочёты

+++ src/test/java/com/vasche/repository/LineRepositoryTest.java
@@ -4,6 +4,7 @@ import com.vasche.entity.Hall;
 import com.vasche.entity.Line;
 import com.vasche.entity.Seat;
 import com.vasche.exception.RepositoryException;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -35,22 +36,30 @@ public class LineRepositoryTest extends RepositoryTestBase {
     void save() throws RepositoryException {
         Optional<Hall> hall = hallRepository.findByName(HALL_NAME1);
 
-        Line line = getLine(1, hall.get().getId());
-        Line actualResult = lineRepository.save(line);
+        if (hall.isPresent()) {
+            Line line = getLine(1, hall.get().getId());
+            Line actualResult = lineRepository.save(line);
 
-        assertThat(actualResult).isNotNull();
+            assertThat(actualResult).isNotNull();
+        } else {
+            Assertions.fail();
+        }
     }
 
     @Test
     void findById() throws RepositoryException {
         Optional<Hall> hall = hallRepository.findByName(HALL_NAME1);
 
-        Line line = lineRepository.save(getLine(1, hall.get().getId()));
+        if (hall.isPresent()) {
+            Line line = lineRepository.save(getLine(1, hall.get().getId()));
 
-        Optional<Line> actualResult = lineRepository.findById(line.getId());
+            Optional<Line> actualResult = lineRepository.findById(line.getId());
 
-        assertThat(actualResult).isPresent();
-        assertThat(actualResult.get()).isEqualTo(line);
+            assertThat(actualResult).isPresent();
+            assertThat(actualResult.get()).isEqualTo(line);
+        } else {
+            Assertions.fail();
+        }
     }
 
     @Test
@@ -59,37 +68,38 @@ public class LineRepositoryTest extends RepositoryTestBase {
         Optional<Hall> hall1 = hallRepository.findByName(HALL_NAME1);
         Optional<Hall> hall2 = hallRepository.findByName(HALL_NAME2);
 
-        Line line1 = lineRepository.save(getLine(1, hall1.get().getId()));
-        Line line2 = lineRepository.save(getLine(2, hall2.get().getId()));
-        Line line3 = lineRepository.save(getLine(1, hall1.get().getId()));
+        if (hall1.isPresent() && hall2.isPresent()) {
+            Line line1 = lineRepository.save(getLine(1, hall1.get().getId()));
+            Line line2 = lineRepository.save(getLine(2, hall2.get().getId()));
+            Line line3 = lineRepository.save(getLine(1, hall1.get().getId()));
 
-        List<Line> actualResult = lineRepository.findAll();
+            List<Line> actualResult = lineRepository.findAll();
 
-        assertThat(actualResult).hasSize(3);
-        List<Integer> moviesIds = actualResult.stream()
-                .map(Line::getId)
-                .toList();
-        assertThat(moviesIds).contains(line1.getId(), line2.getId(), line3.getId());
-
+            assertThat(actualResult).hasSize(3);
+            List<Integer> moviesIds = actualResult.stream().map(Line::getId).toList();
+            assertThat(moviesIds).contains(line1.getId(), line2.getId(), line3.getId());
+        } else {
+            Assertions.fail();
+        }
     }
 
     @Test
     void findAllByHallId() throws RepositoryException {
 
         Optional<Hall> hall1 = hallRepository.findByName(HALL_NAME1);
-        Optional<Hall> hall2 = hallRepository.findByName(HALL_NAME2);
 
-        Line line1 = lineRepository.save(getLine(1, hall1.get().getId()));
-        Line line2 = lineRepository.save(getLine(2, hall2.get().getId()));
-        Line line3 = lineRepository.save(getLine(1, hall1.get().getId()));
+        if (hall1.isPresent()) {
+            Line line1 = lineRepository.save(getLine(1, hall1.get().getId()));
+            Line line3 = lineRepository.save(getLine(1, hall1.get().getId()));
 
-        List<Line> actualResult = lineRepository.findAllByHallId(hall1.get().getId());
+            List<Line> actualResult = lineRepository.findAllByHallId(hall1.get().getId());
 
-        assertThat(actualResult).hasSize(2);
-        List<Integer> moviesIds = actualResult.stream()
-                .map(Line::getId)
-                .toList();
-        assertThat(moviesIds).contains(line1.getId(), line3.getId());
+            assertThat(actualResult).hasSize(2);
+            List<Integer> moviesIds = actualResult.stream().map(Line::getId).toList();
+            assertThat(moviesIds).contains(line1.getId(), line3.getId());
+        } else {
+            Assertions.fail();
+        }
 
     }
 
@@ -98,14 +108,18 @@ public class LineRepositoryTest extends RepositoryTestBase {
 
         Optional<Hall> hall = hallRepository.findByName(HALL_NAME1);
 
-        Line line = lineRepository.save(getLine(1, hall.get().getId()));
+        if (hall.isPresent()) {
+            Line line = lineRepository.save(getLine(1, hall.get().getId()));
 
-        Seat seat = seatRepository.save(getSeat(1, line.getId()));
+            Seat seat = seatRepository.save(getSeat(1, line.getId()));
 
-        Optional<Line> actualResult = lineRepository.findBySeatId(seat.getId());
+            Optional<Line> actualResult = lineRepository.findBySeatId(seat.getId());
 
-        assertThat(actualResult).isPresent();
-        assertThat(actualResult.get()).isEqualTo(line);
+            assertThat(actualResult).isPresent();
+            assertThat(actualResult.get()).isEqualTo(line);
+        } else {
+            Assertions.fail();
+        }
 
     }
 
@@ -122,33 +136,51 @@ public class LineRepositoryTest extends RepositoryTestBase {
     void deleteExistingEntity() throws RepositoryException {
         Optional<Hall> hall = hallRepository.findByName(HALL_NAME1);
 
-        Line line = lineRepository.save(getLine(1, hall.get().getId()));
+        if (hall.isPresent()) {
+            Line line = lineRepository.save(getLine(1, hall.get().getId()));
 
-        boolean actualResult = lineRepository.delete(line.getId());
+            boolean actualResult = lineRepository.delete(line.getId());
 
-        assertTrue(actualResult);
+            assertTrue(actualResult);
+        } else {
+            Assertions.fail();
+        }
+
     }
 
     @Test
     void deleteNotExistingEntity() throws RepositoryException {
         Optional<Hall> hall = hallRepository.findByName(HALL_NAME1);
-        lineRepository.save(getLine(1, hall.get().getId()));
 
-        boolean actualResult = lineRepository.delete(100500);
-        assertFalse(actualResult);
+        if (hall.isPresent()) {
+            lineRepository.save(getLine(1, hall.get().getId()));
+
+            boolean actualResult = lineRepository.delete(100500);
+            assertFalse(actualResult);
+        } else {
+            Assertions.fail();
+        }
     }
 
     @Test
     void update() throws RepositoryException {
         Optional<Hall> hall = hallRepository.findByName(HALL_NAME1);
 
-        Line line = lineRepository.save(getLine(1, hall.get().getId()));
+        if (hall.isPresent()) {
+            Line line = lineRepository.save(getLine(1, hall.get().getId()));
 
-        line.setNumber(2);
+            line.setNumber(2);
 
-        lineRepository.update(line);
+            lineRepository.update(line);
 
-        Line updatedLine = lineRepository.findById(line.getId()).get();
-        assertThat(updatedLine).isEqualTo(line);
+            if (lineRepository.findById(line.getId()).isPresent()) {
+                Line updatedLine = lineRepository.findById(line.getId()).get();
+                assertThat(updatedLine).isEqualTo(line);
+            } else {
+                Assertions.fail();
+            }
+        } else {
+            Assertions.fail();
+        }
     }
 }
